Fixed Warning - in messages_send_boot_req.
Introdotto un cast per evitare un warning in un confronto tra size_t e ssize_t.

+++ messages.c
@@ -111,7 +111,7 @@ int messages_send_boot_req(int sockfd, const struct sockaddr* dest, socklen_t de
         return -1;
 
     sendLen = sendto(sockfd, bootmsg, msgLen, 0, dest, destLen);
-    if (sendLen != msgLen)
+    if ((size_t)sendLen != msgLen)
         return -1;
 
     return 0;
